requirements.txt
requests
mysql-connector-python
fastapi
uvicorn
streamlit
aiohttp
openai